Move sort files.
