+++ sort/sort_test.go
@@ -1,4 +1,4 @@
-package slice
+package sort
 
 import (
 	"sort"
